simplyblock_web/requirements.txt
flask
py-cpuinfo
kubernetes